Rename the CLI from newrelic-lambda-layers-cli to newrelic-lambda-cli (#7)
* Rename the CLI from newrelic-lambda-layers-cli to newrelic-lambda-cli

See: https://newrelic.atlassian.net/browse/LAMBDA-165

* Update author_email
