+++ .pre-commit-config.yaml
@@ -1,0 +1,5 @@
+-   repo: https://github.com/psf/black
+    rev: stable
+    hooks:
+    - id: black
+      language_version: python3.7
