+++ MANIFEST.in
@@ -1,2 +1,2 @@
 include *.md LICENSE
-recursive-include newrelic_lambda_layers *.py *.yaml
+recursive-include newrelic_lambda_cli *.py *.yaml
