+++ newrelic_lambda_cli/templates/newrelic-log-ingestion.yaml
