+++ newrelic_lambda_cli/templates/nr-lambda-integration-role.yaml
