+++ newrelic_lambda_cli/utils.py
@@ -1,3 +1,5 @@
+from __future__ import absolute_import
+
 import sys
 
 import boto3
