--- newrelic_lambda_layers/layers.py
@@ -1,9 +1,0 @@
-import requests
-
-
-def index(region, runtime):
-    req = requests.get(
-        f"https://{region}.nr-layers.iopipe.com/get-layers?CompatibleRuntime={runtime}"
-    )
-    layers_response = req.json()
-    return layers_response.get("Layers", [])
